Move logging unsupported fields into logger package
